requirements.txt
# ── Core framework ─────────────────────────────────────────────────
flask==3.1.0
flask-sqlalchemy==3.1.1
flask-login==0.6.3
werkzeug>=3.0

# ── Environment / config ──────────────────────────────────────────
python-dotenv==1.0.1

# ── HTTP client ───────────────────────────────────────────────────
requests>=2.31

# ── Analytics (optional — fails silently if not configured) ───────
posthog>=3.3

# ── Production server + database ─────────────────────────────────
gunicorn>=21.2
psycopg2-binary>=2.9
